+++ R/gd_optunidisc.R
@@ -1,17 +1,13 @@
-#' @title best univariate discretization based on geodetector q-statistic
+#' @title optimal univariate discretization based on geodetector q-statistic
 #' @author Wenbo Lv \email{lyu.geosocial@gmail.com}
-#' @description
-#' Function for determining the best univariate discretization based on geodetector q-statistic.
 #'
-#' @param formula A formula of best univariate discretization.
-#' @param data A `data.frame` or `tibble` of observation data.
-#' @param discnum (optional) A vector of number of classes for discretization. Default is `3:8`.
+#' @param formula A formula.
+#' @param data A `data.frame`, `tibble` or `sf` object of observation data.
+#' @param discnum (optional) A vector of numbers of discretization. Default is `3:8`.
 #' @param discmethod (optional) A vector of methods for discretization, default is using
 #' `c("sd","equal","geometric","quantile","natural")` by invoking `sdsfun`.
 #' @param cores (optional) Positive integer (default is 1). When cores are greater than 1, use
 #' multi-core parallel computing.
-#' @param return_disc (optional) Whether or not return discretized result used the optimal parameter.
-#' Default is `TRUE`.
 #' @param seed (optional) Random seed number, default is `123456789`.
 #' @param ... (optional) Other arguments passed to `sdsfun::discretize_vector()`.
 #'
@@ -20,36 +16,34 @@
 #' \item{\code{x}}{the name of the variable that needs to be discretized}
 #' \item{\code{k}}{optimal discretization number}
 #' \item{\code{method}}{optimal discretization method}
+#' \item{\code{qstatistic}}{optimal q-statistic}
 #' \item{\code{disc}}{optimal discretization results}
 #' }
 #' @export
 #'
 #' @examples
 #' data('sim')
-#' gd_bestunidisc(y ~ xa + xb + xc,
-#'                data = sim,
-#'                discnum = 3:6)
+#' gd_optunidisc(y ~ xa + xb + xc,
+#'               data = sim,
+#'               discnum = 3:6)
 #'
-gd_bestunidisc = \(formula, data, discnum = 3:8,
-                   discmethod = c("sd","equal","geometric","quantile","natural"),
-                   cores = 1, return_disc = TRUE, seed = 123456789, ...){
+gd_optunidisc = \(formula, data, discnum = 3:8,
+                  discmethod = c("sd","equal","geometric","quantile","natural"),
+                  cores = 1, seed = 123456789, ...){
   doclust = FALSE
-  if (inherits(cores, "cluster")) {
+  if (cores > 1) {
     doclust = TRUE
-  } else if (cores > 1) {
-    doclust = TRUE
-    cores = parallel::makeCluster(cores)
-    on.exit(parallel::stopCluster(cores), add=TRUE)
+    cl = parallel::makeCluster(cores)
+    on.exit(parallel::stopCluster(cl), add=TRUE)
   }
 
-  formula = stats::as.formula(formula)
-  formula.vars = all.vars(formula)
-  response = data[, formula.vars[1], drop = TRUE]
-  if (formula.vars[2] == "."){
-    explanatory = data[,-which(colnames(data) == formula.vars[1])]
-  } else {
-    explanatory = subset(data, TRUE, match(formula.vars[-1], colnames(data)))
+  if (inherits(data,"sf")){
+    data = sf::st_drop_geometry(data)
   }
+
+  formulavars = sdsfun::formula_varname(formula,data)
+  response = data[, formulavars[[1]], drop = TRUE]
+  explanatory = data[, formulavars[[2]]]
 
   discname = names(explanatory)
   paradf = tidyr::crossing("x" = discname,
@@ -67,7 +61,7 @@ gd_bestunidisc = \(formula, data, discnum = 3:8,
   }
 
   if (doclust) {
-    out_g = parallel::parLapply(cores,parak,calcul_disc)
+    out_g = parallel::parLapply(cl,parak,calcul_disc)
     out_g = tibble::as_tibble(do.call(rbind, out_g))
   } else {
     out_g = purrr::map_dfr(parak,calcul_disc)
@@ -78,16 +72,15 @@ gd_bestunidisc = \(formula, data, discnum = 3:8,
     dplyr::slice_max(order_by = qstatistic,
                      with_ties = FALSE) %>%
     dplyr::ungroup() %>%
-    dplyr::select(-qstatistic) %>%
+    dplyr::select(dplyr::everything(),qstatistic) %>%
     as.list()
 
-  if(return_disc){
-    suppressMessages({resdisc = purrr::pmap_dfc(out_g,
-                                \(x,k,method) sdsfun::discretize_vector(
-                                x = explanatory[,x,drop = TRUE],
-                                n = k, method = method, ...)) %>%
-      purrr::set_names(out_g[[1]])})
-    out_g = append(out_g,list("disv" = resdisc))
-  }
+  suppressMessages({resdisc = purrr::pmap_dfc(out_g[1:3],
+                                              \(x,k,method) sdsfun::discretize_vector(
+                                                x = explanatory[,x,drop = TRUE],
+                                                n = k, method = method, seed = seed, ...)) %>%
+    purrr::set_names(out_g[[1]])})
+  out_g = append(out_g,list("disc" = resdisc))
+
   return(out_g)
 }