+++ R/idsa.R
@@ -1,7 +1,5 @@
 #' @title interactive detector for spatial associations(IDSA) model
 #' @author Wenbo Lv \email{lyu.geosocial@gmail.com}
-#' @description
-#' Function for interactive detector for spatial associations model.
 #' @references
 #' Yongze Song & Peng Wu (2021) An interactive detector for spatial associations,
 #' International Journal of Geographical Information Science, 35:8, 1676-1701,
@@ -19,8 +17,7 @@
 #' @param wt (optional) The spatial weight matrix. When `data` is not an `sf` object, must provide `wt`.
 #' @param discnum (optional) Number of multilevel discretization. Default will use `3:8`.
 #' @param discmethod (optional) The discretization methods. Default all use `quantile`.
-#' Noted that `robust` will use `robust_disc()`; `rpart` will use `rpart_disc()`;
-#' Others use `sdsfun::discretize_vector()`.
+#' Noted that  `rpart` will use `rpart_disc()`; Others use `sdsfun::discretize_vector()`.
 #' @param overlay (optional) Spatial overlay method. One of `and`, `or`, `intersection`.
 #' Default is `and`.
 #' @param strategy (optional) Discretization strategy. When `strategy` is `1L`, choose the highest SPADE model q-statistics to