+++ R/opgd.R
@@ -1,7 +1,5 @@
 #' @title optimal parameters-based geographical detector(OPGD) model
 #' @author Wenbo Lv \email{lyu.geosocial@gmail.com}
-#' @description
-#' Function for optimal parameters-based geographical detector(OPGD) model.
 #' @references
 #' Song, Y., Wang, J., Ge, Y. & Xu, C. (2020) An optimal parameters-based geographical detector
 #' model enhances geographic characteristics of explanatory variables for spatial heterogeneity
@@ -21,8 +19,8 @@
 #' @param type (optional) The type of geographical detector, which must be `factor`(default),
 #' `interaction`, `risk`, `ecological`. You can run one or more types at one time.
 #' @param alpha (optional) Specifies the size of confidence level. Default is `0.95`.
-#' @param ... (optional) Other arguments passed to `gd_bestunidisc()`. A useful parameter is `seed`,
-#'  which is used to set the random number seed.
+#' @param ... (optional) Other arguments passed to `gd_optunidisc()`. A useful parameter is `seed`,
+#' which is used to set the random number seed.
 #'
 #' @return A list.
 #' \describe{
@@ -55,15 +53,15 @@ opgd = \(formula, data, discvar = NULL, discnum = 3:8,
     discvar = colnames(data)[-which(colnames(data) == yname)]
   }
   discdf = dplyr::select(data,dplyr::all_of(c(yname,discvar)))
-  g = gd_bestunidisc(paste0(yname,'~',paste0(discvar,collapse = '+')),
-                     data = discdf, discnum = discnum,
-                     discmethod = discmethod, cores = cores, ...)
+  g = gd_optunidisc(paste0(yname,'~',paste0(discvar,collapse = '+')),
+                    data = discdf, discnum = discnum,
+                    discmethod = discmethod, cores = cores, ...)
   opt_param = tibble::as_tibble(g[1:3])
   names(opt_param) = c("varibale","discnum","method")
   discedvar = colnames(data[,-which(colnames(data) %in% discvar)])
   newdata = data %>%
     dplyr::select(dplyr::all_of(discedvar)) %>%
-    dplyr::bind_cols(g$disv)
+    dplyr::bind_cols(g$disc)
   if (length(type) == 1){
     res = gd(paste0(yname,' ~ .'),data = newdata,type = type,alpha = alpha)
   } else {