+++ R/zzz.R
@@ -1,5 +1,10 @@
 .onLoad = function(...) {
   loadNamespace("Rcpp")
+  loadNamespace("sf")
   loadNamespace("tibble")
-  loadNamespace("magrittr")
+}
+
+.onAttach = function(...){
+  packageStartupMessage("gdverse ", utils::packageVersion("gdverse"))
+  packageStartupMessage("See itmsa package for entropy-based spatial association measures.")
 }